+++ src/main/java/com/mercadolibre/planning/model/api/web/controller/request/HeadcountProductivityDataRequest.java
@@ -12,5 +12,5 @@ public class HeadcountProductivityDataRequest {
     @NotNull
     private OffsetTime dayTime;
 
-    private double productivity;
+    private long productivity;
 }
